services/products/requirements.txt
Django
djangorestframework
djongo
psycopg2
